+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## v3.4.0-a2 - 2024-02-02
+
+### Changes:
+
+* 'Post to Discord' button always visible (but greyed out if not usable) on Fleet Carrier and CMDR Information windows.
+
+### Bug Fixes:
+
+* If one of the discord webhook URLs was invalid, was failing immediately and wasn't trying any additional webhooks.
+* Discord button enabling / disabling wasn't working properly with the new multiple discord channel support.
+
+
 ## v3.4.0-a1 - 2024-01-28
 
 ### New Features: